introduce internal getVariableSignature utility
now extracted also to graphql-js PR, see https://github.com/graphql/graphql-js/pull/4175
